# chore: Optimize the use of deepcopy in Translator._get_function_names()